1. added an args.py file to parse user arguments; 2. update machine_config.py for better naming of directories created (avoid creating repetitive Data/ directory); 3. update pre_processing.py to extract features with foundation models UNI, prov-gigapath, resnet-50 and also compute weighted adjacency matrices as in CAMIL
